refactor io_utils
